+++ lib/packages/wikibase/data-model/tests/unit/Entity/ItemIdSetTest.php
@@ -16,6 +16,27 @@ use Wikibase\DataModel\Entity\ItemIdSet;
  */
 class ItemIdSetTest extends \PHPUnit_Framework_TestCase {
 
+	public function testGetIterator() {
+		$set = new ItemIdSet();
+		$this->assertInstanceOf( 'Traversable', $set->getIterator() );
+	}
+
+	/**
+	 * @dataProvider serializationsProvider
+	 */
+	public function testGetSerializations( array $itemIds, array $expected ) {
+		$set = new ItemIdSet( $itemIds );
+		$this->assertEquals( $expected, $set->getSerializations() );
+	}
+
+	public function serializationsProvider() {
+		return array(
+			array( array(), array() ),
+			array( array( new ItemId( 'Q1' ) ), array( 'Q1' ) ),
+			array( array( new ItemId( 'Q1' ), new ItemId( 'Q2' ) ), array( 'Q1', 'Q2' ) ),
+		);
+	}
+
 	public function testGivenEmptySet_countReturnsZero() {
 		$set = new ItemIdSet();
 		$this->assertEquals( 0, $set->count() );